Move Clipboard in a own modul.
