+++ opengever/base/browser/copy_items.py
@@ -1,40 +1,10 @@
 from five import grok
-from OFS.CopySupport import cookie_path
 from opengever.base import _
+from opengever.base.clipboard import Clipboard
 from plone import api
-from plone.app.uuid.utils import uuidToObject
 from plone.dexterity.interfaces import IDexterityContainer
-from plone.uuid.interfaces import IUUID
 from plone.z3cform import layout
 from Products.statusmessages.interfaces import IStatusMessage
-import json
-
-
-class Clipboard(object):
-
-    key = '_clipboard'
-
-    def __init__(self, request):
-        self.request = request
-
-    def set_objs(self, objs):
-        uuids = [IUUID(obj) for obj in objs]
-        value = json.dumps(uuids).encode('base64')
-        self.request.RESPONSE.setCookie(
-            self.key, value, path=str(cookie_path(self.request)))
-
-    def get_objs(self):
-        value = self.request.cookies.get(self.key)
-        if value:
-            uuids = json.loads(value.decode('base64'))
-            return [uuidToObject(uuid) for uuid in uuids]
-
-        return None
-
-    def cookie_path(self):
-        # Return a "path" value for use in a cookie that refers
-        # to the root of the Zope object space.
-        return self.request['BASEPATH1'] or "/"
 
 
 class CopyItemsFormView(layout.FormWrapper, grok.View):
